Imported from iceweasel_14.0.1.orig.tar.bz2

+++ dom/sms/src/ril/SmsDatabaseService.js
@@ -7,11 +7,12 @@
 const {classes: Cc, interfaces: Ci, utils: Cu, results: Cr} = Components;
 
 Cu.import("resource://gre/modules/XPCOMUtils.jsm");
+Cu.import("resource://gre/modules/Services.jsm");
 
 const RIL_SMSDATABASESERVICE_CONTRACTID = "@mozilla.org/sms/rilsmsdatabaseservice;1";
 const RIL_SMSDATABASESERVICE_CID = Components.ID("{a1fa610c-eb6c-4ac2-878f-b005d5e89249}");
 
-const DEBUG = true;
+const DEBUG = false;
 const DB_NAME = "sms";
 const DB_VERSION = 1;
 const STORE_NAME = "sms";
@@ -22,6 +23,11 @@ const DELIVERY_RECEIVED = "received";
 const FILTER_TIMESTAMP = "timestamp";
 const FILTER_NUMBERS = "numbers";
 const FILTER_DELIVERY = "delivery";
+
+const READ_ONLY = "readonly";
+const READ_WRITE = "readwrite";
+const PREV = "prev";
+const NEXT = "next";
 
 XPCOMUtils.defineLazyServiceGetter(this, "gSmsService",
                                    "@mozilla.org/sms/smsservice;1",
@@ -41,16 +47,20 @@ const GLOBAL_SCOPE = this;
  * SmsDatabaseService
  */
 function SmsDatabaseService() {
+  // Prime the directory service's cache to ensure that the ProfD entry exists
+  // by the time IndexedDB queries for it off the main thread. (See bug 743635.)
+  Services.dirsvc.get("ProfD", Ci.nsIFile);
+
   gIDBManager.initWindowless(GLOBAL_SCOPE);
 
   let that = this;
-  this.newTxn(Ci.nsIIDBTransaction.READ_ONLY, function(error, txn, store){
+  this.newTxn(READ_ONLY, function(error, txn, store){
     if (error) {
       return;
     }
     // In order to get the highest key value, we open a key cursor in reverse
     // order and get only the first pointed value.
-    let request = store.openCursor(null, Ci.nsIIDBCursor.PREV);
+    let request = store.openCursor(null, PREV);
     request.onsuccess = function onsuccess(event) {
       let cursor = event.target.result;
       if (!cursor) {
@@ -165,7 +175,7 @@ SmsDatabaseService.prototype = {
    * Start a new transaction.
    *
    * @param txn_type
-   *        Type of transaction (e.g. IDBTransaction.READ_WRITE)
+   *        Type of transaction (e.g. READ_WRITE)
    * @param callback
    *        Function to call when the transaction is available. It will
    *        be invoked with the transaction and the 'sms' object store.
@@ -254,7 +264,7 @@ SmsDatabaseService.prototype = {
   onMessageListCreated: function onMessageListCreated(messageList, requestId) {
     if (DEBUG) debug("Message list created: " + messageList);
     let self = this;
-    self.newTxn(Ci.nsIIDBTransaction.READ_ONLY, function (error, txn, store) {
+    self.newTxn(READ_ONLY, function (error, txn, store) {
       if (error) {
         gSmsRequestManager.notifyReadMessageListFailed(
           requestId, Ci.nsISmsRequestManager.INTERNAL_ERROR);
@@ -295,7 +305,7 @@ SmsDatabaseService.prototype = {
     this.lastKey += 1;
     message.id = this.lastKey;
     if (DEBUG) debug("Going to store " + JSON.stringify(message));
-    this.newTxn(Ci.nsIIDBTransaction.READ_WRITE, function(error, txn, store) {
+    this.newTxn(READ_WRITE, function(error, txn, store) {
       if (error) {
         return;
       }
@@ -313,7 +323,7 @@ SmsDatabaseService.prototype = {
   saveReceivedMessage: function saveReceivedMessage(sender, body, date) {
     let message = {delivery:  DELIVERY_RECEIVED,
                    sender:    sender,
-                   receiver:  null,  //TODO see bug 733266
+                   receiver:  this.mRIL.radioState.msisdn, 
                    body:      body,
                    timestamp: date};
     return this.saveMessage(message);
@@ -321,7 +331,7 @@ SmsDatabaseService.prototype = {
 
   saveSentMessage: function saveSentMessage(receiver, body, date) {
     let message = {delivery:  DELIVERY_SENT,
-                   sender:    null, //TODO see bug 733266
+                   sender:    this.mRIL.radioState.msisdn,
                    receiver:  receiver,
                    body:      body,
                    timestamp: date};
@@ -330,7 +340,7 @@ SmsDatabaseService.prototype = {
 
   getMessage: function getMessage(messageId, requestId) {
     if (DEBUG) debug("Retrieving message with ID " + messageId);
-    this.newTxn(Ci.nsIIDBTransaction.READ_ONLY, function (error, txn, store) {
+    this.newTxn(READ_ONLY, function (error, txn, store) {
       if (error) {
         if (DEBUG) debug(error);
         gSmsRequestManager.notifyGetSmsFailed(
@@ -382,46 +392,28 @@ SmsDatabaseService.prototype = {
   },
 
   deleteMessage: function deleteMessage(messageId, requestId) {
+    let deleted = false;
     let self = this;
-    this.newTxn(Ci.nsIIDBTransaction.READ_WRITE, function (error, txn, store) {
+    this.newTxn(READ_WRITE, function (error, txn, store) {
       if (error) {
         gSmsRequestManager.notifySmsDeleteFailed(
           requestId, Ci.nsISmsRequestManager.INTERNAL_ERROR);
         return;
       }
-      let request = store.delete(messageId);
+      let request = store.count(messageId);
 
-      request.onerror = function onerror(event) {
-        if (DEBUG) debug("Caught error on request ", event.target.errorCode);
-        //TODO look at event.target.errorCode
-        gSmsRequestManager.notifySmsDeleteFailed(
-          requestId, Ci.nsISmsRequestManager.INTERNAL_ERROR);
+      request.onsuccess = function onsuccess(event) {        
+        let count = event.target.result;
+        if (DEBUG) debug("Count for messageId " + messageId + ": " + count);
+        deleted = (count == 1);
+        if (deleted) {
+          store.delete(messageId);
+        }
       };
 
       txn.oncomplete = function oncomplete(event) {
         if (DEBUG) debug("Transaction " + txn + " completed.");
-        // Once we transaction is done, we need to check if we actually deleted
-        // the message. As IndexedDB does not provide the affected records info,
-        // we need to try to get the message from the database again to check
-        // that it is actually gone.
-        self.newTxn(Ci.nsIIDBTransaction.READ_ONLY, function (error, txn, store) {
-          let request = store.getAll(messageId);
-          request.onsuccess = function onsuccess(event) {
-            let deleted = (event.target.result.length == 0);
-            gSmsRequestManager.notifySmsDeleted(requestId, deleted);
-          };
-          request.onerror = function onerror(event) {
-            if (DEBUG) {
-              debug("Error checking the message deletion " +
-                    event.target.errorCode);
-            }
-            //TODO should we notify here as an internal error? The failed check
-            //     does not mean that the deletion has failed, so maybe we
-            //     should notify successfully.
-            gSmsRequestManager.notifySmsDeleteFailed(
-              requestId, Ci.nsISmsRequestManager.INTERNAL_ERROR);
-          };
-        });
+        gSmsRequestManager.notifySmsDeleted(requestId, deleted);
       };
 
       txn.onerror = function onerror(event) {
@@ -478,7 +470,7 @@ SmsDatabaseService.prototype = {
     };
 
     let self = this;
-    this.newTxn(Ci.nsIIDBTransaction.READ_ONLY, function (error, txn, store) {
+    this.newTxn(READ_ONLY, function (error, txn, store) {
       if (error) {
         errorCb(error);
         return;
@@ -495,7 +487,7 @@ SmsDatabaseService.prototype = {
       } else if (filter.endDate != null) {
         timeKeyRange = IDBKeyRange.upperBound(filter.endDate.getTime());
       }
-      let direction = reverse ? Ci.nsIIDBCursor.PREV : Ci.nsIIDBCursor.NEXT;
+      let direction = reverse ? PREV : NEXT;
       let timeRequest = store.index("timestamp").openKeyCursor(timeKeyRange,
                                                                direction);
 
@@ -573,7 +565,7 @@ SmsDatabaseService.prototype = {
       gSmsRequestManager.notifyNoMessageInList(requestId);
       return;
     }
-    this.newTxn(Ci.nsIIDBTransaction.READ_ONLY, function (error, txn, store) {
+    this.newTxn(READ_ONLY, function (error, txn, store) {
       if (DEBUG) debug("Fetching message " + messageId);
       let request = store.get(messageId);
       let message;
@@ -615,6 +607,12 @@ SmsDatabaseService.prototype = {
   }
 
 };
+
+XPCOMUtils.defineLazyGetter(SmsDatabaseService.prototype, "mRIL", function () {
+    return Cc["@mozilla.org/telephony/system-worker-manager;1"]
+              .getService(Ci.nsIInterfaceRequestor)
+              .getInterface(Ci.nsIRadioInterfaceLayer);
+});
 
 const NSGetFactory = XPCOMUtils.generateNSGetFactory([SmsDatabaseService]);
